Bugzilla/requirements.txt
mkdocs
python-dotenv
pytest
ruff
-e .
great_expectations
deepchecks[tabular]
anywidget
pyyaml==6.0.1

fastapi
ijson

gradio>=4.0.0

uvicorn

numpy>=1.23,<1.26
pandas>=2.2
scikit-learn>=1.4
dvc[s3]>=3.50
dagshub>=0.3

mlflow==2.16.2
tensorflow-cpu==2.15.0
keras==2.15.*
tensorflow_probability==0.23.0
matplotlib

setfit
alibi-detect==0.12.0
sentence-transformers==3.0.1
torch==2.3.1
datasets
prometheus-client==0.21.0
